+++ index.html
@@ -97,6 +97,6 @@
       </div>
     </main>
 
-    <script type="module" src="js/index.js"></script>
+    <script type="module" src="./index.js"></script>
   </body>
 </html>
