--- js/state/regist.js
@@ -1,9 +1,0 @@
-import { dateStore } from "./state.js";
-
-//렌더링
-// dataStore.subscribe('year', (year) => { 
-    // const yearEl = dateStore.getState('year');
-
-// }
-
-
