+++ setting/config.js
